+++ app/src/main/java/com/kontranik/easycycle/models/CDay.kt
@@ -5,6 +5,6 @@ import java.util.*
 class CDay (
     var id: Long,
     var date: Date,
-    var cyclesDay: Long,
+    var cyclesDay: Int,
     var phases: List<Phase>,
     var color: String?) {}
